# service/src/main/resources/api-spec.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/testC":{"get":{"tags":["c-controller"],"operationId":"testGet","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}},"put":{"tags":["c-controller"],"operationId":"testPut","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["c-controller"],"operationId":"testPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}}},"/testB":{"get":{"tags":["b-controller"],"operationId":"testGet_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}},"put":{"tags":["b-controller"],"operationId":"testPut_1","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["b-controller"],"operationId":"testPost_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}}},"/testA":{"get":{"tags":["a-controller"],"operationId":"testGet_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}},"put":{"tags":["a-controller"],"operationId":"testPut_2","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["a-controller"],"operationId":"testPost_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RequestDto"}}}}}}}},"components":{"schemas":{"RequestDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zip":{"type":"string"},"country":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"},"created":{"type":"string","format":"date-time"},"updated":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"mdmId":{"type":"string","format":"uuid"},"countOfUnsuccessfulLogin":{"type":"integer","format":"int32"},"amount":{"type":"number"}}}}}}